refactor: Get요청일 때, Get요청을 확인하는 로직 추가
