+++ src/main/java/webserver/request/Request.java
@@ -70,6 +70,10 @@ public class Request {
         }
     }
 
+    public boolean isGetRequest() {
+        return line.isGet();
+    }
+
     // line 이 GET이면 URI를 획득, POST면 URI를 획득하지 못함.
     public Optional<String> getUri() {
         return line.getUri();
